Implement fast events experiment
Three sliders.

Two of them are controlled. When value changes it is send through events
port to Elm program and when new state comes from state port they are
both updated. Note that this is not how sliders are typically used.
Typically they are uncontrolled.

Third one is uncontrolled - it is there for reference.

The performance of the managed ones is visibly worse, but IMO within
reasonable expectations.

Possible solutions: debounce the update, only update on